refactor(validation): use zod schemas for POST /accounts validation (#72)
rather than AJV.

Also marks AJV validation function as deprecated and urges the use of zod.

Note that this PR drops support for throwing a validation error when an unsupported fiat account type is posted. Bundling that in with validation was nifty, though a bit more opinionated. I'm open to ideas on how to preserve that functionality, but it might just be an inherent tradeoff of importing zod schemas from fiatconnect-types, which cannot know which particular fiat account schemas will be supported by a given provider.